requirements.txt
crewai
crewai-tools
python-dotenv
streamlit
pydantic
google-generativeai